% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_enetwild_data.R
\name{list_enetwild_data}
\alias{list_enetwild_data}
\title{List available species and years for ENETWILD data}
\usage{
list_enetwild_data()
}
\value{
A list containing the available species and years as character vectors.
}
\description{
This function lists the available species and years for which data is
available in the "inst/extdata" folder of the package. The function loops through
the species folders in the "inst/extdata" directory, and identifies the years for
which data is available by looking at the CSV files in each species folder.
The function then displays the list of species and years in the console
and returns a list of the available species and years.
}
\examples{
list <- list_enetwild_data()
}
