% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_enetwild_data.R
\name{load_enetwild_data}
\alias{load_enetwild_data}
\title{Load data for a specific species and year}
\usage{
load_enetwild_data(species, year)
}
\arguments{
\item{species}{A character string indicating the name of the species to load.}

\item{year}{An integer indicating the year of the data to load.}
}
\value{
A data.frame containing the data for the specified species and year.
}
\description{
This function loads the data for a specific species and year from the
"inst/extdata" folder in the package directory. The function combines the species
name and year to create the file name, and then loads the corresponding CSV
file into R using the "read.csv()" function.
}
\examples{
data <- load_enetwild_data("Sus.scrofa", 2015)
}
